style: :lipstick: added dialog boxes and UI changes
