+++ lib/view/new_invoice_screen/new_client_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:invoice_maker/core/constants/global_key.dart';
-import 'package:invoice_maker/core/utils/coming_soon_dialog.dart';
+import 'package:invoice_maker/core/utils/dialog_boxes.dart';
 import 'package:invoice_maker/core/utils/switch_button.dart';
 import 'package:invoice_maker/core/utils/text_form_validator.dart';
 import 'package:provider/provider.dart';
